strategy as interface that knows about seasons
